+++ SubEthaEdit-Mac/Source/TCMFoundation/NSRangeTCMAdditions.h
@@ -13,7 +13,7 @@ FOUNDATION_STATIC_INLINE BOOL TouchingRanges(NSRange range1, NSRange range2) {
             && NSMaxRange(range1) >= range2.location);
 }
 
-FOUNDATION_STATIC_INLINE int EndCharacterIndex(NSRange aRange) {
+FOUNDATION_STATIC_INLINE NSUInteger EndCharacterIndex(NSRange aRange) {
     if (aRange.length==0) return aRange.location;
     return NSMaxRange(aRange)-1;
 }
